#2 : detach access to tweet table logic
