+++ app/service/twitter_batch_service.py
@@ -3,9 +3,9 @@ from __future__ import annotations
 from datetime import datetime
 from datetime import timedelta
 
-from app.db.twitter import insert_tweet
+from app.db.tweet import insert_tweet
+from app.db.tweet import is_exist_tweet
 from app.db.twitter import insert_tweets_users
-from app.db.twitter import is_exist_tweet
 from app.db.twitter import is_exist_tweets_users
 from app.db.user_oauth_info import get_twitter_all_users_oauth_info
 from app.utils.twitter_api.twitter_timeline import fetch_all_users_timeline_tweets
